requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
pyspark
kafka-python
hdfs